Remove all Matrix allocations and convert parameterized versions.
This removes all allocations of temporary objects from Matrix code
paths, replacing them all with explicit per-thread Context types.
It removes all redundant declarations from the parameterized versions
of matrices (where the extra phantom type parameters on methods could
not convey any useful information).

References #2

+++ io7m-jtensors-core/src/main/java/com/io7m/jtensors/VectorM3F.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 <code@io7m.com> http://io7m.com
+ * Copyright © 2015 <code@io7m.com> http://io7m.com
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -583,7 +583,7 @@ public final class VectorM3F
     final VectorReadable3FType v0,
     final VectorReadable3FType v1)
   {
-    return VectorM3F.magnitude(VectorM3F.subtract(v0, v1, c.v2a));
+    return (double) VectorM3F.magnitude(VectorM3F.subtract(v0, v1, c.v2a));
   }
 
   /**
